refactor: clean up code and remove unnecessary components
- Refactored and optimized codebase by removing obsolete components
- Adjusted and refined elements in the footer for better alignment and presentation
