+++ src/app/(main-routes)/page.tsx
@@ -1,13 +1,10 @@
 import { BecomeATrainer } from '@/ui/modules/become-a-trainer/become-a-trainer';
-// import { Recommandations } from '@/ui/modules/recommandations/recommandations'
-import { HowItWorks } from '@/ui/modules/how-it-works/how-it-works';
 import { HeroBanner } from '@/ui/modules/hero-banner/hero-banner';
-// import { Benefits } from '@/ui/modules/benefits/benefits'
 import { Newsletter } from '@/ui/modules/newsletter/newsletter';
 import { Container } from '@/ui/components/container/container';
 import { BenefitSearch } from '@/components/benefitSeach';
 import { Recommandations } from '@/ui/modules/recommandations/recommandations';
-import { HowItWork } from '@/ui/modules/HowItWork';
+import { HowItWork } from '@/ui/modules/how-it-works/HowItWork';
 
 export default function Home() {
   return (
@@ -18,7 +15,6 @@ export default function Home() {
         <Recommandations />
         <BecomeATrainer />
         <HowItWork />
-        {/* <HowItWorks /> */}
         <Newsletter />
       </Container>
     </main>
